requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
chromadb==0.4.18
llama-index==0.9.13
groq==0.4.0
pymupdf==1.23.8
python-docx==1.1.0